added semi-colon after return statement, added closing quote to post request argument.

+++ AssociateEvaluationSystem/src/main/java/com/revature/aes/grading/AssessmentGrader.java
@@ -225,7 +225,7 @@ public class AssessmentGrader {
 			log.info(" keyFileName: " + keyFileName);
 			itemWeight = templateDataMap.get(key).getWeight();
 			log.info(" weight: "+itemWeight);
-			int codeTestResult = sec.evaluateSnippet(userFileName, keyFileName);
+			double codeTestResult = sec.evaluateSnippet(userFileName, keyFileName);
 			log.info(" code being evaluated to "+codeTestResult);
 			itemWeightedGrade = itemWeight*codeTestResult;//
 			result[0] = result[0]+itemWeightedGrade;
